pi/room_node/requirements.txt
numpy>=1.24
scipy>=1.10
paho-mqtt>=1.6
rns>=0.7
